How this works:
This project is a game manager that recreates the method of modern game consoles. It take login information to allow the user to access the program, gives the user 3 games of choice, allows them to switch from each without restarting, and provides a place to play games in one area.